+++ flask_security/datastore.py
@@ -300,7 +300,14 @@ class UserDatastore(object):
            be stored directly in the DB. Do NOT pass in a plaintext password!
            Best practice is to pass in ``hash_password(plaintext_password)``.
 
-        The new user's ``active`` property will be set to true.
+           Furthermore, no validation is done on the password (e.g for minimum length).
+           Best practice is to call
+           ``app.security._password_validator(plaintext_password, True)``
+           and look for a ``None`` return meaning the password conforms to the
+           configured validations.
+
+        The new user's ``active`` property will be set to ``True``
+        unless explicitly set to ``False`` in `kwargs`.
         """
         kwargs = self._prepare_create_user_args(**kwargs)
         user = self.user_model(**kwargs)