+++ tests/test_csrf.py
@@ -268,8 +268,8 @@ def test_cp_with_token(app, client):
     # make sure returned csrf_token works in header.
     data = dict(
         password="password",
-        new_password="newpassword",
-        new_password_confirm="newpassword",
+        new_password="battery staple",
+        new_password_confirm="battery staple",
     )
 
     with mp_validate_csrf() as mp:
@@ -353,8 +353,8 @@ def test_different_mechanisms(app, client):
         # session based change password should fail
         data = dict(
             password="password",
-            new_password="newpassword",
-            new_password_confirm="newpassword",
+            new_password="battery staple",
+            new_password_confirm="battery staple",
         )
 
         response = client.post(
@@ -396,8 +396,8 @@ def test_different_mechanisms_nc(app, client_nc):
         # token based should work
         data = dict(
             password="password",
-            new_password="newpassword",
-            new_password_confirm="newpassword",
+            new_password="battery staple",
+            new_password_confirm="battery staple",
         )
         response = client_nc.post(
             "/change",
@@ -445,8 +445,8 @@ def test_cp_with_token_cookie(app, client):
     # make sure returned csrf_token works in header.
     data = dict(
         password="password",
-        new_password="newpassword",
-        new_password_confirm="newpassword",
+        new_password="battery staple",
+        new_password_confirm="battery staple",
     )
     csrf_token = [c.value for c in client.cookie_jar if c.name == "X-XSRF-Token"][0]
     with mp_validate_csrf() as mp:
@@ -480,8 +480,8 @@ def test_cp_with_token_cookie_expire(app, client):
     time.sleep(2)
     data = dict(
         password="password",
-        new_password="newpassword",
-        new_password_confirm="newpassword",
+        new_password="battery staple",
+        new_password_confirm="battery staple",
     )
     csrf_token = [c.value for c in client.cookie_jar if c.name == "X-XSRF-Token"][0]
     with mp_validate_csrf() as mp:
@@ -519,8 +519,8 @@ def test_cp_with_token_cookie_refresh(app, client):
     # make sure returned csrf_token works in header.
     data = dict(
         password="password",
-        new_password="newpassword",
-        new_password_confirm="newpassword",
+        new_password="battery staple",
+        new_password_confirm="battery staple",
     )
 
     csrf_cookie = [c for c in client.cookie_jar if c.name == "X-XSRF-Token"][0]